+++ Blurred-ios/ChannelViewController.swift
@@ -35,6 +35,8 @@ class ChannelViewController: UIViewController, UINavigationControllerDelegate, U
                 let json = try JSONSerialization.jsonObject(with: data!, options: .mutableContainers) as? NSDictionary
                 if let parseJSON = json {
                     let imageUrl: String? = parseJSON["thumbnail_url"] as? String
+                    let videoUrl: String? = parseJSON["video_url"] as? String
+                    self.videoUrlString = videoUrl!
                     let railsUrl = URL(string: "http://10.0.0.2:3000\(imageUrl!)")
                     DispatchQueue.main.async {
                         Nuke.loadImage(with: railsUrl!, into: cell.thumbnailView)
@@ -55,11 +57,28 @@ class ChannelViewController: UIViewController, UINavigationControllerDelegate, U
         task.resume()
         return cell
     }
-    
+    func seeVideo() {
+        self.performSegue(withIdentifier: "showVideo", sender: self)
+    }
+    func collectionView(CollectionView: UICollectionView, didSelectRowAt indexPath: IndexPath) {
+        let destinationVC = ChannelVideoViewController()
+        destinationVC.performSegue(withIdentifier: "showVideo", sender: self)
+    }
+    var videoUrlString = String()
     @IBOutlet weak var collectionView: UICollectionView!
-    
-    
-
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?)
+    {
+        if segue.destination is ChannelVideoViewController
+        {
+            if let vc = segue.destination as? ChannelVideoViewController {
+                if segue.identifier == "showVideo" {
+                    vc.videoString = videoUrlString
+                }
+            } else {
+                self.showErrorContactingServer()
+            }
+        }
+    }
     @IBOutlet weak var followersLabel: UILabel!
     @IBOutlet weak var followingLabel: UILabel!
     @IBOutlet weak var usernameLabel: UILabel!