+++ CHANGELOG.md
@@ -6,12 +6,13 @@
 * Added thumbnail view of video in upload details controller.
 * Added back button to upload details.
 * Added channel video controller for videos on channels.
-* Added collection view to channel view controller (doesn't do anything yet)
+* Added collection view to channel view controller
 * Added collection view to other channel view controller (doesn't do anything yet)
 * Moved profile divider line down by 5px
 * Added thumbnails to videos on your own channel
 * Updated api request for avatar urls.
 * Added thumbnails to videos on others' channels 
+* Video playback controller for channel videos (Can only play 1st video of user.)
 ## v0.2.5 (PRERELEASE)
 * Account settings editing now works (actually contacts the api)
 * Fixed bug where timer would stop forever and load any user info until app restart